remove power_density and excl_area from plexos_nodes update tests
