+++ reVX/plexos/rev_reeds_plexos.py
@@ -4,277 +4,23 @@ Created on Wed Aug 21 13:47:43 2019
 
 @author: gbuster
 """
-import os
 from concurrent.futures import as_completed
 import json
+import logging
 import numpy as np
+import os
 import pandas as pd
 from scipy.spatial import cKDTree
 from warnings import warn
-import logging
 
 from rex.utilities.execution import SpawnProcessPool
 
-from reVX.plexos.utilities import parse_table_name
 from reVX.handlers.outputs import Outputs
-from reVX.plexos.utilities import DataCleaner, get_coord_labels
-
+from reVX.plexos.plexos_node import PlexosNode
+from reVX.plexos.utilities import (DataCleaner, get_coord_labels,
+                                   parse_table_name)
 
 logger = logging.getLogger(__name__)
-
-
-class PlexosNode:
-    """Framework to analyze the gen profile at a single plexos node."""
-
-    def __init__(self, sc_build, cf_fpath, cf_res_gids, power_density,
-                 exclusion_area=0.0081, forecast_fpath=None,
-                 forecast_map=None):
-        """
-        Parameters
-        ----------
-        sc_build : pd.DataFrame
-            Supply curve buildout table. Must only have rows that are built
-            in this plexos node. Must have resource_gid lookup, counts per
-            resource_gid, and capacity at each SC point.
-        cf_fpath : str
-            File path to capacity factor file to get profiles from.
-        cf_res_gids : list | np.ndarray
-            Resource GID's available in cf_fpath.
-        power_density : float
-            Power density associated with the current buildout.
-        exclusion_area : float
-            Area in km2 associated with a single exclusion pixel.
-        forecast_fpath : str | None
-            Forecasted capacity factor .h5 file path (reV results).
-            If not None, the generation profiles are sourced from this file.
-        forecast_map : np.ndarray | None
-            (n, 1) array of forecast meta data indices mapped to the generation
-            meta indices where n is the number of generation points. None if
-            no forecast data being considered.
-        """
-        self._sc_build = sc_build
-        self._cf_fpath = cf_fpath
-        self._cf_res_gids = cf_res_gids
-        self._power_density = power_density
-        self._exclusion_area = exclusion_area
-        self._forecast_fpath = forecast_fpath
-        self._forecast_map = forecast_map
-
-    def _get_sc_meta(self, row_idx):
-        """Get meta for SC point row index, which is part of this plexos node.
-
-        Parameters
-        ----------
-        row_idx : int
-            Index value for the row of the target SC point in self._sc_build.
-
-        Returns
-        -------
-        sc_meta : pd.DataFrame
-            Dataframe with rows corresponding to resource/generation pixels
-            that are part of this SC point. Sorted by cf_mean with best
-            cf_mean at top.
-        """
-
-        res_gids, gid_counts, _ = self._parse_sc_point(row_idx)
-
-        gid_capacity = [self._power_density * self._exclusion_area * c
-                        for c in gid_counts]
-        gen_gids = [np.where(self._cf_res_gids == g)[0][0]
-                    for g in res_gids]
-        sc_meta = pd.DataFrame({'gen_gids': gen_gids,
-                                'res_gids': res_gids,
-                                'gid_counts': gid_counts,
-                                'gid_capacity': gid_capacity})
-        sc_meta = sc_meta.sort_values(by='gen_gids')
-
-        with Outputs(self._cf_fpath, mode='r') as cf_outs:
-            cf_mean = cf_outs['cf_mean', list(sc_meta['gen_gids'].values)]
-
-        sc_meta['cf_mean'] = cf_mean
-        sc_meta = sc_meta.sort_values(by='cf_mean', ascending=False)
-        sc_meta = sc_meta.reset_index(drop=True)
-
-        # infinite capacity in the last gid to make sure full buildout is done
-        sc_meta.loc[sc_meta.index[-1], 'gid_capacity'] = 1e6
-
-        return sc_meta
-
-    def _parse_sc_point(self, row_idx):
-        """Parse data from sc point.
-
-        Parameters
-        ----------
-        row_idx : int
-            Index value for the row of the target SC point in self._sc_build.
-
-        Returns
-        -------
-        res_gids : list
-            Resource GIDs associated with SC point i.
-        gid_counts : list
-            Number of exclusion pixels that are not excluded associated
-            with each res_gid.
-        buildout : float
-            Total REEDS requested buildout associated with SC point i.
-        """
-
-        buildout = float(self._sc_build.loc[row_idx, 'built_capacity'])
-
-        res_gids = self._sc_build.loc[row_idx, 'res_gids']
-        gid_counts = self._sc_build.loc[row_idx, 'gid_counts']
-
-        if isinstance(res_gids, str):
-            res_gids = json.loads(res_gids)
-
-        if isinstance(gid_counts, str):
-            gid_counts = json.loads(gid_counts)
-
-        return res_gids, gid_counts, buildout
-
-    def _build_sc_profile(self, row_idx, profile):
-        """Build a power generation profile based on SC point i.
-
-        Parameters
-        ----------
-        row_idx : int
-            Index value for the row of the target SC point in self._sc_build.
-        profile : np.ndarray | None
-            (t,) array of generation in MW, or None if this is the first
-            SC point to add generation.
-
-        Returns
-        ----------
-        profile : np.ndarray
-            (t,) array of generation in MW where t is the timeindex length.
-        res_gids : list
-            List of resource GID's that were built from this SC point.
-        gen_gids : list
-            List of generation GID's that were built from this SC point.
-        res_built : list
-            List of built capacities at each resource GID from this SC point.
-        """
-
-        sc_meta = self._get_sc_meta(row_idx)
-        _, _, buildout = self._parse_sc_point(row_idx)
-
-        res_gids = []
-        gen_gids = []
-        res_built = []
-
-        for row_jdx in sc_meta.index.values:
-
-            if buildout <= sc_meta.loc[row_jdx, 'gid_capacity']:
-                to_build = buildout
-            else:
-                to_build = sc_meta.loc[row_jdx, 'gid_capacity']
-
-            buildout -= to_build
-
-            res_built.append(np.round(to_build, decimals=2))
-
-            if self._forecast_map is None:
-                gen_gid = sc_meta.loc[row_jdx, 'gen_gids']
-                with Outputs(self._cf_fpath, mode='r') as cf_outs:
-                    cf_profile = cf_outs['cf_profile', :, gen_gid]
-            else:
-                gen_gid = self._forecast_map[sc_meta.loc[row_jdx, 'gen_gids']]
-                with Outputs(self._forecast_fpath, mode='r') as cf_outs:
-                    cf_profile = cf_outs['cf_profile', :, gen_gid]
-
-            res_gids.append(sc_meta.loc[row_jdx, 'res_gids'])
-            gen_gids.append(gen_gid)
-
-            if profile is None:
-                profile = to_build * cf_profile
-            else:
-                profile += to_build * cf_profile
-
-            if buildout <= 0:
-                break
-
-        if len(profile.shape) != 1:
-            profile = profile.flatten()
-
-        return profile, res_gids, gen_gids, res_built
-
-    def _make_node_profile(self):
-        """Make an aggregated generation profile for a single plexos node.
-
-        Returns
-        -------
-        profile : np.ndarray
-            (t, ) array of generation in MW.
-        res_gids : list
-            List of resource GID's that were built for this plexos node.
-        gen_gids : list
-            List of generation GID's that were built for this plexos node.
-        res_built : list
-            List of built capacities at each resource GID for this plexos node.
-        """
-
-        profile = None
-        res_gids = []
-        gen_gids = []
-        res_built = []
-
-        for i in self._sc_build.index.values:
-
-            profile, i_res_gids, i_gen_gids, i_res_built = \
-                self._build_sc_profile(i, profile)
-
-            res_gids += i_res_gids
-            gen_gids += i_gen_gids
-            res_built += i_res_built
-
-        return profile, res_gids, gen_gids, res_built
-
-    @classmethod
-    def run(cls, sc_build, cf_fpath, cf_res_gids, power_density,
-            exclusion_area=0.0081, forecast_fpath=None, forecast_map=None):
-        """Make an aggregated generation profile for a single plexos node.
-
-        Parameters
-        ----------
-        sc_build : pd.DataFrame
-            Supply curve buildout table. Must only have rows that are built
-            in this plexos node. Must have resource_gid lookup, counts per
-            resource_gid, and capacity at each SC point.
-        cf_fpath : str
-            File path to capacity factor file to get profiles from.
-        cf_res_gids : list | np.ndarray
-            Resource GID's available in cf_fpath.
-        power_density : float
-            Power density associated with the current buildout.
-        exclusion_area : float
-            Area in km2 associated with a single exclusion pixel.
-        forecast_fpath : str | None
-            Forecasted capacity factor .h5 file path (reV results).
-            If not None, the generation profiles are sourced from this file.
-        forecast_map : np.ndarray | None
-            (n, 1) array of forecast meta data indices mapped to the generation
-            meta indices where n is the number of generation points. None if
-            no forecast data being considered.
-
-        Returns
-        -------
-        profile : np.ndarray
-            (t, ) array of generation in MW.
-        res_gids : list
-            List of resource GID's that were built for this plexos node.
-        gen_gids : list
-            List of generation GID's that were built for this plexos node.
-        res_built : list
-            List of built capacities at each resource GID for this plexos node.
-        """
-
-        n = cls(sc_build, cf_fpath, cf_res_gids, power_density,
-                exclusion_area=exclusion_area, forecast_fpath=forecast_fpath,
-                forecast_map=forecast_map)
-
-        profile, res_gids, gen_gids, res_built = n._make_node_profile()
-
-        return profile, res_gids, gen_gids, res_built
 
 
 class PlexosAggregation:
@@ -283,8 +29,7 @@ class PlexosAggregation:
     """
 
     def __init__(self, plexos_nodes, rev_sc, reeds_build, cf_fpath,
-                 forecast_fpath=None, build_year=2050, exclusion_area=0.0081,
-                 max_workers=None):
+                 forecast_fpath=None, build_year=2050, max_workers=None):
         """
         Parameters
         ----------
@@ -302,19 +47,14 @@ class PlexosAggregation:
             If not None, the generation profiles are sourced from this file.
         build_year : int
             REEDS year of interest.
-        exclusion_area : float
-            Area in km2 of an exclusion pixel.
         max_workers : int | None
             Do node aggregation on max_workers.
         """
-
         self._plexos_nodes = DataCleaner.rename_cols(plexos_nodes)
         self._cf_fpath = cf_fpath
         self._forecast_fpath = forecast_fpath
         self.build_year = build_year
-        self.exclusion_area = exclusion_area
         self._cf_res_gids = None
-        self._power_density = None
         self._plexos_meta = None
         self._time_index = None
         self.max_workers = max_workers
@@ -434,23 +174,6 @@ class PlexosAggregation:
 
         return self._cf_res_gids
 
-    @property
-    def power_density(self):
-        """Get the mean power density based on the reV SC capacity and area.
-
-        Returns
-        -------
-        power_density : float
-            Estimated power density based on (capacity / area).
-        """
-
-        if self._power_density is None:
-            pd = (self._sc_build['potential_capacity'].values
-                  / self._sc_build['area_sq_km'].values)
-            self._power_density = np.round(np.mean(pd))
-
-        return self._power_density
-
     def _check_gids(self):
         """Ensure that the SC buildout GIDs are available in the cf file.
 
@@ -460,7 +183,6 @@ class PlexosAggregation:
             List of missing supply curve gids
             (in reeds but not in reV resource).
         """
-
         bad_sc_points = []
         missing = list(set(self.sc_res_gids) - set(self.available_res_gids))
         if any(missing):
@@ -780,8 +502,7 @@ class PlexosAggregation:
                 mask = (self._node_map == inode)
                 f = exe.submit(PlexosNode.run,
                                self._sc_build[mask], self._cf_fpath,
-                               self.available_res_gids, self.power_density,
-                               exclusion_area=self.exclusion_area,
+                               cf_res_gids=self.available_res_gids,
                                forecast_fpath=self._forecast_fpath,
                                forecast_map=self._forecast_map)
                 futures[f] = i
@@ -815,8 +536,7 @@ class PlexosAggregation:
             mask = (self._node_map == inode)
             p = PlexosNode.run(
                 self._sc_build[mask], self._cf_fpath,
-                self.available_res_gids, self.power_density,
-                exclusion_area=self.exclusion_area,
+                cf_res_gids=self.available_res_gids,
                 forecast_fpath=self._forecast_fpath,
                 forecast_map=self._forecast_map)
 
@@ -835,8 +555,7 @@ class PlexosAggregation:
 
     @classmethod
     def run(cls, plexos_nodes, rev_sc, reeds_build, cf_fpath,
-            forecast_fpath=None, build_year=2050, exclusion_area=0.0081,
-            max_workers=None):
+            forecast_fpath=None, build_year=2050, max_workers=None):
         """Run plexos aggregation.
 
         Parameters
@@ -855,8 +574,6 @@ class PlexosAggregation:
             If not None, the generation profiles are sourced from this file.
         build_year : int
             REEDS year of interest.
-        exclusion_area : float
-            Area in km2 of an exclusion pixel.
         max_workers : int | None
             Do node aggregation on max_workers.
 
@@ -872,7 +589,7 @@ class PlexosAggregation:
 
         pa = cls(plexos_nodes, rev_sc, reeds_build, cf_fpath,
                  forecast_fpath=forecast_fpath, build_year=build_year,
-                 exclusion_area=exclusion_area, max_workers=max_workers)
+                 max_workers=max_workers)
         profiles = pa._make_profiles()
 
         return pa.plexos_meta, pa.time_index, profiles
