+++ tests/data/aggregated_plexos_profiles/plexos_meta.csv
@@ -1,4 +1,4 @@
 ,sc_gid,plexos_id,latitude,longitude,voltage,interconnect,built_capacity,res_gids,gen_gids,res_built
-5696,37900,70937_BLANCAPEAK,37.6097398,-105.86891570000002,115,WI,2.794500097632411,"[827348, 827348]","[1361, 1361]","[0.3, 2.49]"
-11090,43735,70459_WALSENBG,37.61254,-104.77489,230,WI,22.45319931944661,"[839089, 839090, 838017, 839084, 838012, 839084, 838012, 839084, 839081, 838011, 839078, 839079, 839078]","[2017, 2018, 1939, 2012, 1934, 2012, 1934, 2012, 2009, 1933, 2006, 2007, 2006]","[0.66, 1.56, 3.21, 0.83, 6.4, 0.83, 7.0, 0.27, 0.36, 1.02, 0.02, 0.28, 0.02]"
-11339,43997,12091_YORKCANY,36.851310999999995,-104.913054,115,WI,4.957199329604211,"[837998, 837996]","[1922, 1920]","[0.43, 4.53]"
+5696,37900,70937_BLANCAPEAK,37.6097398,-105.86891570000002,115,WI,2.794500097632411,"[827348, 827349, 826284, 828413, 826285, 828412, 826286, 827350, 826287, 828414, 829477, 829478, 827351, 828415, 829476, 829479, 827348, 826284, 828412, 826283, 827347, 829476, 828411, 826282, 827346, 829475, 828410, 829474, 829473, 828409, 827345, 826281]","[1361, 1362, 1327, 1403, 1328, 1402, 1329, 1363, 1330, 1404, 1444, 1445, 1364, 1405, 1443, 1446, 1361, 1327, 1402, 1326, 1360, 1443, 1401, 1325, 1359, 1442, 1400, 1441, 1440, 1399, 1358, 1324]","[0.01631, 0.03849, 0.01001, 0.03849, 0.02736, 0.01758, 0.02699, 0.03878, 0.00779, 0.03819, 0.00957, 0.00994, 0.00979, 0.00816, 0.00445, 0.00185, 0.18, 0.14533, 0.17574, 0.22986, 0.31317, 0.03953, 0.32411, 0.2329, 0.31317, 0.07297, 0.31986, 0.07297, 0.00912, 0.02676, 0.02554, 0.00973]"
+11090,43735,70459_WALSENBG,37.61254,-104.77489,230,WI,22.45319931944661,"[839089, 839090, 838017, 838018, 839091, 839092, 838019, 838016, 836947, 836946, 836945, 836944, 835876, 835875, 835874, 835873, 839084, 838012, 839085, 838011, 836940, 839086, 836941, 838013, 835870, 836939, 835869, 835871, 835868, 839084, 838012, 839085, 838011, 836940, 839086, 836941, 838013, 835870, 836939, 835869, 835871, 835868, 839084, 839081, 838011, 839083, 839082, 838008, 838009, 838010, 836939, 836937, 836936, 836938, 835866, 835865, 835868, 835867, 839078, 839079, 839080, 838006, 839081, 838007, 838008, 836934, 836935, 836936, 835863, 835865, 835864, 839078, 839076, 839077, 838005, 838004, 838003, 836934, 836933, 835863, 836932, 836931, 835862, 835860, 835861]","[2017, 2018, 1939, 1940, 2019, 2020, 1941, 1938, 1859, 1858, 1857, 1856, 1771, 1770, 1769, 1768, 2012, 1934, 2013, 1933, 1852, 2014, 1853, 1935, 1765, 1851, 1764, 1766, 1763, 2012, 1934, 2013, 1933, 1852, 2014, 1853, 1935, 1765, 1851, 1764, 1766, 1763, 2012, 2009, 1933, 2011, 2010, 1930, 1931, 1932, 1851, 1849, 1848, 1850, 1761, 1760, 1763, 1762, 2006, 2007, 2008, 1928, 2009, 1929, 1930, 1846, 1847, 1848, 1758, 1760, 1759, 2006, 2004, 2005, 1927, 1926, 1925, 1846, 1845, 1758, 1844, 1843, 1757, 1755, 1756]","[0.03572, 0.08467, 0.68398, 0.68133, 0.08996, 0.02778, 0.19448, 0.24607, 0.23284, 0.68927, 0.69324, 0.23284, 0.21432, 0.57152, 0.57417, 0.16669, 0.06001, 0.91072, 0.08472, 0.60538, 0.91248, 0.08119, 0.90895, 0.91778, 0.79952, 0.60714, 0.80482, 0.03353, 0.50301, 0.06495, 0.98572, 0.09169, 0.65523, 0.98763, 0.08787, 0.98381, 0.99336, 0.86537, 0.65714, 0.8711, 0.0363, 0.54444, 0.00444, 0.00605, 0.06898, 0.01654, 0.01452, 0.06858, 0.20857, 0.20736, 0.07504, 0.21139, 0.06213, 0.21462, 0.18678, 0.05365, 0.06898, 0.18477, 7e-05, 0.00178, 0.00193, 0.03864, 0.00111, 0.03878, 0.0261, 0.03663, 0.03819, 0.02685, 0.03382, 0.02484, 0.035, 0.00014, 9e-05, 0.00012, 0.00307, 0.00305, 0.00201, 0.00012, 0.00306, 0.00012, 0.00309, 0.00194, 0.00288, 0.0017, 0.00289]"
+11339,43997,12091_YORKCANY,36.851310999999995,-104.913054,115,WI,4.957199329604211,"[837998, 837999, 838000, 836926, 836927, 836928, 835856, 835857, 835855, 837996, 837995, 837997, 836923, 836924, 836925, 836926, 835852, 835853, 835855, 835854]","[1922, 1923, 1924, 1840, 1841, 1842, 1753, 1754, 1752, 1920, 1919, 1921, 1837, 1838, 1839, 1840, 1749, 1750, 1752, 1751]","[0.05357, 0.05409, 0.03634, 0.05129, 0.05336, 0.03738, 0.05212, 0.03914, 0.04796, 0.56649, 0.3618, 0.56981, 0.36512, 0.57866, 0.57645, 0.02434, 0.32529, 0.55986, 0.04647, 0.55764]"
